+++ R/polarAnnulus.R
@@ -153,7 +153,6 @@ polarAnnulus <-
     k = c(20, 10),
     normalise = FALSE,
     breaks = NULL,
-    labels = NULL,
     key.title = paste(statistic, pollutant, sep = " "),
     key.position = "right",
     auto.text = TRUE,
@@ -211,6 +210,9 @@ polarAnnulus <-
     extra.args$title <- quickText(extra.args$title, auto.text)
     extra.args$subtitle <- quickText(extra.args$subtitle, auto.text)
     extra.args$tag <- quickText(extra.args$tag, auto.text)
+
+    # deal with breaks
+    break_opts <- resolve_break_opts(breaks, extra.args)
 
     # separate handling for being overwritten
     if ("caption" %in% names(extra.args)) {
@@ -561,11 +563,10 @@ polarAnnulus <-
     key.title <- check_key_header(key.title, extra.args)
 
     # handle breaks
-    categorical <- !is.null(breaks)
+    categorical <- !is.null(break_opts$breaks)
     results.grid$pred <- cut_plot_breaks(
       results.grid$pred,
-      breaks = breaks,
-      labels = labels
+      break_opts
     )
 
     # plotting