+++ R/utilities.R
@@ -204,29 +204,6 @@ map_type <- function(
   return(out)
 }
 
-#' Create nice labels out of breaks, if only breaks are provided
-#' @noRd
-get_labels_from_breaks <- function(breaks, labels = NULL, sep = " - ") {
-  if (is.null(labels) || anyNA(labels)) {
-    labels <- paste(
-      format(
-        utils::head(breaks, -1),
-        scientific = FALSE,
-        trim = TRUE,
-        drop0trailing = TRUE
-      ),
-      format(
-        utils::tail(breaks, -1),
-        scientific = FALSE,
-        trim = TRUE,
-        drop0trailing = TRUE
-      ),
-      sep = sep
-    )
-  }
-  labels
-}
-
 #' pad out a set of numbers with zeroes to create consistent width
 #' @noRd
 pad_string <- function(y, n = NULL) {
@@ -472,45 +449,6 @@ capture_dots <- function(...) {
   extra.args
 }
 
-# Handle cutting numeric vectors for discretising plots
-cut_plot_breaks <- function(x, breaks, labels) {
-  if (is.null(breaks)) {
-    return(x)
-  }
-
-  # if only one break, use as number of intervals
-  if (length(breaks) == 1) {
-    x <- cut_vec_numeric(x = x, type = "", n.levels = breaks, is.axis = FALSE)
-
-    if (!is.null(labels)) {
-      if (length(labels) != nlevels(x)) {
-        cli::cli_abort(
-          "Length of {.arg labels} ({length(labels)}) \\
-          is not equal to number of categories ({nlevels(x)})."
-        )
-      }
-      levels(x) <- labels
-    }
-
-    return(x)
-  }
-
-  # assign labels if no labels are given
-  labels <- labels %||% get_labels_from_breaks(breaks, labels)
-
-  # ensure breaks covers all data
-  if (max(breaks) < max(x, na.rm = TRUE)) {
-    breaks[length(breaks)] <- ceiling(max(x, na.rm = TRUE))
-  }
-  if (min(breaks) > min(x, na.rm = TRUE)) {
-    breaks[1] <- floor(min(x, na.rm = TRUE))
-  }
-
-  # cut x into breaks
-  breaks <- sort(unique(breaks))
-  cut(x, breaks = breaks, labels = labels, include.lowest = TRUE)
-}
-
 # Recycle helper
 recycle_to_length <- function(x, n, expect1 = FALSE) {
   if (length(x) == n) {
@@ -529,6 +467,21 @@ recycle_to_length <- function(x, n, expect1 = FALSE) {
       x <- c(x, x)
     }
     x <- x[seq_len(n)]
+  }
+  x
+}
+
+# Round values to the point in which they're no longer unique
+round_safely <- function(x, n = 10, n_min = 0) {
+  progress <- TRUE
+  while (progress && n >= n_min) {
+    new_x <- round(x, n)
+    if (length(unique(new_x)) != length(unique(x))) {
+      progress <- FALSE
+    } else {
+      x <- new_x
+      n <- n - 1
+    }
   }
   x
 }